move NetworkContent to presentation/common
move NetworkContent to presentation/common
