+++ app/src/test/java/com/babylon/wallet/android/presentation/transfer/TransferViewModelTest.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.SavedStateHandle
 import app.cash.turbine.ReceiveTurbine
 import app.cash.turbine.test
 import com.babylon.wallet.android.data.dapp.IncomingRequestRepository
-import com.babylon.wallet.android.domain.NetworkContent
+import com.babylon.wallet.android.presentation.common.NetworkContent
 import com.babylon.wallet.android.domain.model.assets.AccountWithAssets
 import com.babylon.wallet.android.domain.usecases.GetAccountDepositResourceRulesUseCase
 import com.babylon.wallet.android.domain.usecases.GetNetworkInfoUseCase
